Refactor: give SelfRelativePtr its own header, shared by both arches
`SelfRelativePtr` is a generic representation, not an hbg type detail: it
carries no knowledge of what it points at, and `PTO2TaskSlotState` already
names two different targets through it. It lived at the bottom of
`pto_runtime2_types.h`, in two byte-identical copies — one per arch.

It moves to `src/common/host_build_graph/self_relative_ptr.h`, one copy for
both arches, reached through the `src/common` include root the platform
already puts on every target's path.

The class body is unchanged. Its doc comment drops the enumeration of the two
users, which had to be restated on every new one, and states the precondition
as the obligation it is: a block whose extent can approach the int32 bound has
to reject it up front, which is what `attach_populated` does.

ctest 114/114; a2a3/a5 x sim/onboard host_build_graph all build.
